Added an option for maximum size.
Read from the config file as max_size = size, where size is in a format
accepted by rsync.

This is useful for automatically excluding large binary files, on the
assumption that they should be obtainable (eg iso images etc) and need
not be backed up.

If the option is not provided, there is no exclusion for any size.

Conflicts:
	bin/backup

cherry-picked from 4a1fc2bf9c20c5fca1472212364338b4790b264e.