manage the {daily,weekly,monthly}.0 symlinks better
"ln -sf SRC DEST" doesn't always do what the manpage says.  If DEST exists
and points to a directory, "ln -sf" does not remove DEST, instead a link
named DEST/SRC is created, pointing to SRC.

This can be a problem if the number of snapshots (of some interval)
changes.  If the snapshots are fully populated (so that the ${next}.0
symlink points to an existing directory), then the script would
erroneously create a symlink inside ${next}.0 pointing to the new final
snapshot.