add a new optional config variable: backup_port
This is the port on the remote backup_host that the local backup pusher will
connect to.  Defaults to 22 if omitted.

This config variable lets you run your remote ssh daemon on a non-standard
port.  This may be helpful if you're connecting to it across the
internet, because it makes the backup server a little harder for the
botnets to find.